requirements.txt
selenium==4.35.0
webdriver-manager==4.0.2
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.1.1
pytelegrambotapi==4.29.1
cryptography==41.0.0
